Updated screen 1
Included text fields for start & end locations, a time picker, and a continue button.
Right now, the time picker doesn't have a title associated with it and doesn't come in a DialogFragment (which we should do eventually).
Also, I put the 4 UI elements in a ScrollView, so it works with smaller screens.