Rename button id, create event handler
For the button with the text "view restaurants" (in the status page), I renamed
the button ID and created an event handler. It launches a Google Maps activity.
Unfortunately, launching the activity cancels navigation, so I will have to go
back and change it. I am committing this to have it in the repo.