The MapFragment works
I extended the MapFragment class, so that we can add custom markers and such to
the map. Also, I left the default methods and code in the StatusMapFragment so
that I can look at it more closely later.

Using the subclass was annoying, because the onCreateView method wanted me to
return the "inflated layout" of the fragment, and I didn't want to have to
use/create a custom layout, since we weren't modifying the layout.  I eventually
realized that the layout I return doesn't actually matter (I'm assuming it gets
replaced by Google Maps code). The clean way to solve the problem, however, was
to call the superclass implementation of the method, which I didn't think of
doing until writing this commit message...