Make embeeded map zoom into user's current location
I also started creating the StatusMapFragment correctly (using the built-in
factory method). The way I managed to pass a Context object to the
CenterMapOnCurrentLocation method is a bit janky: The Context object needed is
passed to the onAttach method, which is called before ANYTHING else, Later, the
factory method is called, which passes a NEW StatusMapObject. This method, by
default, doesn't know about the Context object. The factory method is obviously
called from a static context, but I would like each object returned to have its
own reference to a Context object (rather than sharing a static member). The way
I got around this was by saving the Context object passed to onAttach to a
static variable, than assigning it to a non-static member myContext to each
instance of StatusMapFragment before returning it from the factory method.