Update FavRestaurantTypes code
We used to overwrite the SharedPrefs settings with currently-selected "favorite
restaurants". Also, there was a limit of 3 restaurants because the distances
array was of length 3.

We still need to write the currently-selected "favorite restaurants" to the
SharedPrefs so that when the user re-opens the application on navigating mode,
the current restaurants can be reloaded.